+++ testdata/de/dm/intellij/liferay/language/freemarker/LiferayFtlVariableProviderTest/com/liferay/portal/template/ServiceLocator.java
@@ -1,0 +1,9 @@
+package com.liferay.portal.template;
+
+public class ServiceLocator {
+
+    public Object findService(String serviceName) {
+        return null;
+    }
+
+}