+++ changelog.md
@@ -20,6 +20,7 @@ All notable changes to this project will be documented in this file.
 - Feature: Improved support for getSiblings() in Freemarker templates
 - Feature: Inspections for service.xml
 - Feature: Added taglib code completion for class names
+- Feature: Implicit resolve of serviceLocator instructions in Freemarker templates
 
 ## 0.0.2
 - Feature: Added Compass Support for Liferay 6.1 and 6.2